Move Table utils out of the common utils
